segment_anything/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.2.0
future==1.0.0
h11==0.14.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lzstring==1.0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
zipp==3.20.2
